+++ tests/location_tests.js
@@ -389,7 +389,7 @@ describe('openSenseMap API locations tests', function () {
         value: 5.5,
         createdAt: measurement3.createdAt.clone().subtract(2, 'ms')
       };
-      // TODO: something fishy goin on, lands at beginning!!1
+      // WARN: something fishy goin on, lands at beginning!!1
       const measurement1 = {
         value: 5,
         location: [5, 5, 5],
@@ -687,6 +687,8 @@ describe('openSenseMap API locations tests', function () {
     });
 
     it('should provide box.currentLocation for legacy measurements without location field', function () {
+      /* eslint-disable no-console */
+      const originalConsoleLog = console.log;
       /* eslint-disable global-require */
       const { connect, mongoose } = require('../lib/db');
       const Measurement = require('../lib/models/measurement').model;
@@ -694,6 +696,8 @@ describe('openSenseMap API locations tests', function () {
       mongoose.set('debug', false);
 
       // manually create a new measurent without location field
+      console.log = function () {};
+
       return connect()
         .then(function () {
           return new Promise(function (resolve, reject) {
@@ -718,8 +722,11 @@ describe('openSenseMap API locations tests', function () {
           expect(data[0].lon).to.equal('10');
           expect(data[0].lat).to.equal('10');
 
+          console.log = originalConsoleLog;
+
           return chakram.wait();
         });
+      /* eslint-enable no-console */
     });
 
   });
